Two neurons with more testing
I tried to implement the testing function for the delay and some others to check the value of the membrane potential if the external current is positive of negative. I added some constants in the hpp file that will be used for the implementation of 12500 neurons in the whole network. 
Seeing some problems implementing the test function for the new membrane potential I became aware that the function updateState, as the first time of spikes is initialized at 0, was in the refractory state at the beginning so I changed the conditions adding that the spike time should not be 0 and I set the neuron NON_REFRACTORY when I create it. 
For the test functions I implemented the delay one in a way that seems to work and I added to others to check if we put an external current how the membrane potential is going to change.