Comments
In these files I tried to implement the program for 12500 neurons. I had problems with my previous conception of storing the neurons which where sending the spike to one single neuron instead of storing the targets one neuron chooses randomly. 
I chose the second version. 

In the hpp file:
my code has some new attributes(excitatory tells if the neuron is excitatory or inhibitory, targets which stores all the targets of the neuron, spike which is a boolean telling if the neuron spiked). I have implemented the corresponding getters and setters.
A new function, setTargets adds targets to the neuron vector and setRingBuffer increases the ring buffer of J.
The background noise a neuron receives is implemented in the function ExternalSpikes with random number and the Poisson law to get a random distribution. This function is called each time the new membrane potential is calculated.
In the update function, if the neuron spikes the new attribute spike becomes true, it is by default initialized at false. In the update there are two cases for the computation of the membrane potential due to the test to check delay and increment difficult with random values.
The function instaures connections randomly connections between the neurons of the network (parameter). In fillRingBuffer of targets for all of them if the neuron sending the spike is excitatory or inhibitory the value of J is different.

In the main, the vector of all neurons is created with 10000 excitatory and 1250 inhibitory neurons, connections are made and for all steps the neuron is updated. 

For the test, I kept the old ones and declared that they are tests so the old computation of membrane potential should be used.